python/sources/mysql_cdc/requirements.txt
quixstreams==2.9.0
pymysql
mysql-replication
python-dotenv